Remove pragmas, Turn on MSVC Warnings, etc
- Remove every MSVC #pragma warning. MSVC had no warning level set and ran at
  /W1, where almost none of them fired; it now builds our sources at /W3, and
  /WX with SW_WERROR. -Wno-unknown-pragmas goes from the clang/GCC baseline.
- Replace LE_ASSUME with LE_ASSERT. As a release-build hint it changed 88 bytes
  across 696 KB, none in a process path, and a false condition was undefined
  behaviour. Its -Wassume pragmas go with it.
- Make the conversions /W3 reports explicit, and drop a top-level const six
  overrides added that their bases lack (C4373).
- Value-initialise the scale snap's ratios, which GCC 12-16 no longer see
  written without the assume.
- Compute normalisedLinear2dB(double) in double; it was calling the float log10.
- Fold expressions in CompoundChannelState, and drop a stale
  -Wundefined-internal pragma.

What remains is -Wattributes around the CLAP entry and the two -Wuninitialized
sites, which #262 tracks.

Closes #68

+++ src/core/host_interop/programWrite.hpp
@@ -107,7 +107,8 @@ template <class Protocol> class ProgramParameterSetter
         /// \note The destroying overload, deliberately: this is the main thread's
         /// Program, and the main thread is where things are destroyed.
         /// \see AutomatedModuleChain::setParameter.
-        pProgram->moduleChain().setParameter(parameterID.moduleIndex, effectIndex,
+        pProgram->moduleChain().setParameter(parameterID.moduleIndex,
+                                             static_cast<std::int8_t>(effectIndex),
                                              ParametersOnlyModuleInitialiser{});
     }
